backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi==0.116.1",
    "google-auth>=2.40.3",
    "greenlet==3.2.4",
    "griffe==1.11.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "mcp==1.12.4",
    "openai==1.99.1",
    "openai-agents==0.2.5",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pywin32==311",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rpds-py==0.27.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "sse-starlette==3.0.2",
    "starlette==0.47.2",
    "tqdm==4.67.1",
    "types-requests==2.32.4.20250611",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
]
